Fix up serial_interface_output signature

+++ Inc/serial_interface.h
@@ -47,7 +47,7 @@ typedef struct {
 
 void serial_interface_reset();
 void serial_interface_set_output_function(OutputFunction func);
-int serial_interface_output(uint8_t *buffer, uint16_t n);
+uint8_t serial_interface_output(uint8_t *buffer, uint16_t n);
 void serial_interface_consume(uint8_t byte);
 
 #endif  // __INPUT_PROCESSOR__H
