+++ Src/serial_interface.c
@@ -52,7 +52,7 @@ void serial_interface_set_output_function(OutputFunction func) {
 	instance.output_func = func;
 }
 
-int serial_interface_output(uint8_t *buffer, uint16_t n) {
+uint8_t serial_interface_output(uint8_t *buffer, uint16_t n) {
 	return instance.output_func(buffer, n);
 }
